Work around conflict between arydshln and array v2.6
Version 2.6 of array.sty released along with LaTeX2e <2024-06-01>
is not compatible with arydshln's assumption of internal behavior
of array.

Stick to a compatible version of array by adding \RequirePackage{array}
with a rollback option.

Use of date as in [=2024-05-01] makes the LaTeX code compatible
with earlier installation of TeX Live, where the latex engine can pick
a suitable version available.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>