rcuscale: Handle source trees having done allmodconfig builds
When rcuscale is built as a module, that build will generate a file named
kernel/rcu/rcuscale.mod.c, which will cause the defer/rcuscale.sh script
to incorrectly conclude that the kernel source tree is too old for the
script to run.  This commit therefore adjusts the checks to correctly
handle kernel source trees that have done allmodconfig builds.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>