debugging: Move from eqnarray environment to align
The eqnarray environment has served well for decades, but it doesn't
play well with clever references.  Although the only use of eqnarray
doesn't use a \label{}, this commit moves to the align environment.

One thing to keep in mind: the align environment usually has one fewer
instances of "&" per line.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>