api-pthreads/QAfter: Convert timecost.c to clock_gettime()
The timecost.c code uses dgettimeofday(), which is only good down to
a microsecond.  This means that the number of measurements per set must
be long enough to cover enough microseconds to keep the quantization
error down in the noise.  This commit therefore switches to the
nanosecond-granularity clock_gettime() interfaces.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>