CodeSamples: Fix liburcu-signal support deprecation
Gitlab-CI is failing [1] on building CodeSamples due to liburcu v0.15
deprecating [2] support for liburcu-signal:

%%%
   * Deprecation of liburcu-signal

Users can migrate to liburcu-memb with a kernel implementing the
membarrier(2) system call to have similar read-side performance without
requiring use of a reserved signal, and with improved grace period
performance.
%%%

So, in order to avoid breaking builds with liburcu >=0.15 while keeping
support for older systems without sys_membarrier(), check for the existance
of liburcu-signal before adding it to the build command.

Link: https://gitlab.com/linux-kernel/perfbook/-/jobs/8732944799 [1]
Link: https://lists.lttng.org/pipermail/lttng-dev/2024-December/030893.html [2]
Signed-off-by: Leonardo Bras <leobras.c@gmail.com>
Reviewed-by: Akira Yokosawa <akiyks@gmail.com>
Tested-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>