index: List LKMM under "Memory model"
Given that the Linux kernel memory model (LKMM) is a class of "memory
model", it is better be listed under "Memory model" in Index.

As it is currently impossible to make acronym-related terms go to
second level, remove existing index markers of acronym (meaning
\IXacr --> \acr) and use additional index markers of \IXalthmr.

Given it goes under "Memory model", add a redirection of "Linux kernel
consistency model (LKMM)" to "Memory model, Linux kernel" as well.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>