Apply "pdfcrop" to CodeSamples/formal/data/RCU-test-ratio.pdf
RCU-test-ratio.pdf has fairly wide side margins and Figure 12.4 does
not fully utilize available page width.
Crop the plot and include the resulting RCU-test-ratio-crop.pdf so
that it can enjoy wider width.

At the moment, it is the only plot generated on-the-fly, but the
same scheme can be applied in the future to similar plots.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>