Detect buggy fvextra and apply patch on the fly
fvextra versions from v1.11.0 to v1.13.2 (latest) have a minor regression
which ignores fancyvrb's "samepage=true" key.  It is used in perfbook to
prevent short snippets from broken across columns/pages.  This is a minor
regression but affects reproducible build of perfbook.

Unfortunately, distro texlive packages based on upstream TeX Live since
early March 2025 are all affected by this regression, including debian
trixie and opensuse/leap 16.0 (Long Term Support); and ubuntu plucky
and questing (Rolling release); as well as archlinux and
opensuse/tumbleweed (Rapid release) as of this writing.

This issue was reported upstream and a tentative one-liner fix was
provied by Yukai Chou at [1].

Add code in autodate.sh to detect buggy fvextra and have a boolean
"buggyfvextra" set "true" in autodate.tex.

In the preamble, conditionally load fixfvextra.ltx which contains the
tentative patch to be applied on-the-fly.

Link: https://github.com/gpoore/fvextra/issues/32 [1]
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>